+++ .eslintrc.js
@@ -1,8 +1,8 @@
 module.exports = {
   root: true,
   extends: [
-    require.resolve('@vercel/style-guide/eslint/node'),
-    require.resolve('@vercel/style-guide/eslint/typescript'),
+    require.resolve('@khulnasoft/style-guide/eslint/node'),
+    require.resolve('@khulnasoft/style-guide/eslint/typescript'),
     'turbo',
     'prettier',
   ],