+++ packages/remote-nx/src/remote-client.ts
@@ -1,18 +1,18 @@
 import events from 'events';
-import { createClient } from '@vercel/remote';
+import { createClient } from '@khulnasoft/remote';
 import { gray } from 'chalk';
 import type { CustomRunnerOptions } from 'nx-remotecache-custom';
 
-export type VercelRemoteCacheOptions = CustomRunnerOptions<{
+export type KhulnasoftRemoteCacheOptions = CustomRunnerOptions<{
   teamId?: string;
   token?: string;
 }>;
 
-function getCredentials(options: VercelRemoteCacheOptions) {
+function getCredentials(options: KhulnasoftRemoteCacheOptions) {
   // eslint-disable-next-line turbo/no-undeclared-env-vars
-  const token = process.env.VERCEL_ARTIFACTS_TOKEN;
+  const token = process.env.KHULNASOFT_ARTIFACTS_TOKEN;
   // eslint-disable-next-line turbo/no-undeclared-env-vars
-  const teamId = process.env.VERCEL_ARTIFACTS_OWNER;
+  const teamId = process.env.KHULNASOFT_ARTIFACTS_OWNER;
   if (token && teamId) {
     return {
       token,
@@ -22,13 +22,13 @@ function getCredentials(options: VercelRemoteCacheOptions) {
 
   return {
     // eslint-disable-next-line turbo/no-undeclared-env-vars
-    token: process.env.NX_VERCEL_REMOTE_CACHE_TOKEN || options.token,
+    token: process.env.NX_KHULNASOFT_REMOTE_CACHE_TOKEN || options.token,
     // eslint-disable-next-line turbo/no-undeclared-env-vars
-    teamId: process.env.NX_VERCEL_REMOTE_CACHE_TEAM || options.teamId,
+    teamId: process.env.NX_KHULNASOFT_REMOTE_CACHE_TEAM || options.teamId,
   };
 }
 
-export function getVercelRemoteCacheClient(options: VercelRemoteCacheOptions) {
+export function getKhulnasoftRemoteCacheClient(options: KhulnasoftRemoteCacheOptions) {
   // Prevents a client warning of 'MaxListenersExceededWarning'
   // The warning originates when nx-remotecache-custom creates a tarball
   // of the build artifacts.
@@ -38,9 +38,9 @@ export function getVercelRemoteCacheClient(options: VercelRemoteCacheOptions) {
   const { token, teamId } = getCredentials(options);
   if (!token) {
     throw new Error(
-      'Missing a Vercel access token for Vercel Remote Cache. ' +
+      'Missing a Khulnasoft access token for Khulnasoft Remote Cache. ' +
         'Specify a token either in nx.json or using the environment ' +
-        'variable NX_VERCEL_REMOTE_CACHE_TOKEN.',
+        'variable NX_KHULNASOFT_REMOTE_CACHE_TOKEN.',
     );
   }
 
@@ -48,7 +48,7 @@ export function getVercelRemoteCacheClient(options: VercelRemoteCacheOptions) {
     const owner = teamId ? `Team ${teamId}` : 'Personal account';
     // eslint-disable-next-line no-console
     console.log(
-      gray(`Initializing Vercel remote cache with scope: "${owner}".`),
+      gray(`Initializing Khulnasoft remote cache with scope: "${owner}".`),
     );
   }
   return createClient(token, { teamId, product: 'nx' });