+++ packages/remote-nx/src/setup-sdk.ts
@@ -1,14 +1,14 @@
 import { initEnv } from 'nx-remotecache-custom';
-import { getVercelRemoteCacheClient } from './remote-client';
+import { getKhulnasoftRemoteCacheClient } from './remote-client';
 import type { Readable } from 'stream';
-import type { VercelRemoteCacheOptions } from './remote-client';
+import type { KhulnasoftRemoteCacheOptions } from './remote-client';
 
 // eslint-disable-next-line @typescript-eslint/require-await
-const setupSDK = async (options: VercelRemoteCacheOptions) => {
+const setupSDK = async (options: KhulnasoftRemoteCacheOptions) => {
   initEnv(options);
-  const remote = getVercelRemoteCacheClient(options);
+  const remote = getKhulnasoftRemoteCacheClient(options);
   return {
-    name: 'Vercel Remote Cache',
+    name: 'Khulnasoft Remote Cache',
     fileExists: (filename: string) => remote.exists(filename).send(),
     retrieveFile: (filename: string) => remote.get(filename).stream(),
     storeFile: (filename: string, stream: Readable) =>