+++ packages/remote-nx/test/setup-sdk.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, afterEach } from 'vitest';
 import { setupSDK } from '../src/setup-sdk';
-import type { VercelRemoteCacheOptions } from '../src/remote-client';
+import type { KhulnasoftRemoteCacheOptions } from '../src/remote-client';
 
 describe('sdk', () => {
   const originalEnv = process.env;
@@ -10,7 +10,7 @@ describe('sdk', () => {
   });
 
   it('to be defined', async () => {
-    const options: VercelRemoteCacheOptions = {
+    const options: KhulnasoftRemoteCacheOptions = {
       teamId: 'team_id',
       token: 'token',
       lifeCycle: {},
@@ -21,7 +21,7 @@ describe('sdk', () => {
   });
 
   it('to implement interface', async () => {
-    const options: VercelRemoteCacheOptions = {
+    const options: KhulnasoftRemoteCacheOptions = {
       teamId: 'team_id',
       token: 'token',
       lifeCycle: {},
@@ -31,27 +31,27 @@ describe('sdk', () => {
     expect(client.fileExists).toBeInstanceOf(Function);
     expect(client.retrieveFile).toBeInstanceOf(Function);
     expect(client.storeFile).toBeInstanceOf(Function);
-    expect(client.name).toMatch('Vercel Remote Cache');
+    expect(client.name).toMatch('Khulnasoft Remote Cache');
   });
 
   it('to throw error with missing token', async () => {
-    const options: VercelRemoteCacheOptions = {
+    const options: KhulnasoftRemoteCacheOptions = {
       teamId: 'team_id',
       lifeCycle: {},
     };
 
     await expect(setupSDK(options)).rejects.toThrowError(
-      'Missing a Vercel access token',
+      'Missing a Khulnasoft access token',
     );
   });
 
-  it('to allow using NX_VERCEL_REMOTE_CACHE_TOKEN var as token', async () => {
-    const options: VercelRemoteCacheOptions = {
+  it('to allow using NX_KHULNASOFT_REMOTE_CACHE_TOKEN var as token', async () => {
+    const options: KhulnasoftRemoteCacheOptions = {
       teamId: 'team_id',
       lifeCycle: {},
     };
     // eslint-disable-next-line turbo/no-undeclared-env-vars
-    process.env.NX_VERCEL_REMOTE_CACHE_TOKEN = 'token';
+    process.env.NX_KHULNASOFT_REMOTE_CACHE_TOKEN = 'token';
     await expect(setupSDK(options)).resolves.toBeDefined();
   });
 });