+++ packages/remote-nx/tsup.config.ts
@@ -1,10 +1,10 @@
-import { defineConfig } from "tsup";
-import type { Options } from "tsup";
+import { defineConfig } from 'tsup';
+import type { Options } from 'tsup';
 
 // eslint-disable-next-line import/no-default-export
 export default defineConfig((options: Options) => ({
-  entry: ["src/index.ts"],
-  format: ["cjs"],
+  entry: ['src/index.ts'],
+  format: ['cjs'],
   dts: true,
   clean: true,
   ...options,