+++ packages/remote-rush/src/credential.ts
@@ -1,9 +1,9 @@
 import { CredentialCache, EnvironmentConfiguration } from '@rushstack/rush-sdk';
 import type { ICredentialCacheEntry } from '@rushstack/rush-sdk';
 
-const CREDENTIAL_CACHE_PREFIX = 'vercel-remote-cache';
+const CREDENTIAL_CACHE_PREFIX = 'khulnasoft-remote-cache';
 
-interface VercelCredential {
+interface KhulnasoftCredential {
   teamId?: string;
   token?: string;
 }
@@ -16,18 +16,18 @@ function getCacheIds(teamId?: string) {
 }
 
 /**
- * These should only be set during Vercel Deployments
+ * These should only be set during Khulnasoft Deployments
  */
-function getCredentialFromVercelEnv(): VercelCredential {
+function getCredentialFromKhulnasoftEnv(): KhulnasoftCredential {
   return {
     // eslint-disable-next-line turbo/no-undeclared-env-vars
-    token: process.env.VERCEL_ARTIFACTS_TOKEN,
+    token: process.env.KHULNASOFT_ARTIFACTS_TOKEN,
     // eslint-disable-next-line turbo/no-undeclared-env-vars
-    teamId: process.env.VERCEL_ARTIFACTS_OWNER,
+    teamId: process.env.KHULNASOFT_ARTIFACTS_OWNER,
   };
 }
 
-function parseRushCredential(credential: string): VercelCredential {
+function parseRushCredential(credential: string): KhulnasoftCredential {
   const parsed = credential.split(':');
   if (parsed.length === 1) {
     return {
@@ -44,7 +44,7 @@ function parseRushCredential(credential: string): VercelCredential {
 
 async function getCredentialFromUserStore(
   teamId?: string,
-): Promise<VercelCredential> {
+): Promise<KhulnasoftCredential> {
   let cacheEntry: ICredentialCacheEntry | undefined;
   const { globalCacheId, teamCacheId } = getCacheIds(teamId);
 
@@ -73,7 +73,7 @@ async function getCredentialFromUserStore(
   return cacheEntry?.credential ? { token: cacheEntry.credential, teamId } : {};
 }
 
-function getCredentialFromRushEnv(): VercelCredential {
+function getCredentialFromRushEnv(): KhulnasoftCredential {
   const credential = EnvironmentConfiguration.buildCacheCredential;
   if (credential) {
     return parseRushCredential(credential);
@@ -82,10 +82,10 @@ function getCredentialFromRushEnv(): VercelCredential {
 }
 
 /**
- * Tries to find credentials for Vercel Artifacts in the following order:
- * 1. The `VERCEL_ARTIFACTS_TOKEN` and `VERCEL_ARTIFACTS_OWNER` environment variables.
+ * Tries to find credentials for Khulnasoft Artifacts in the following order:
+ * 1. The `KHULNASOFT_ARTIFACTS_TOKEN` and `KHULNASOFT_ARTIFACTS_OWNER` environment variables.
  *
- *    The value of `VERCEL_ARTIFACTS_OWNER` will always override the `teamId` argument.
+ *    The value of `KHULNASOFT_ARTIFACTS_OWNER` will always override the `teamId` argument.
  * 2. The `RUSH_BUILD_CACHE_CREDENTIAL` environment variable which has the format `<token>` or `<teamId>:<token>`.
  *
  *    If the `RUSH_BUILD_CACHE_CREDENTIAL` has the form `<teamId>:<token>`, the provided `<teamId>` will override the
@@ -97,14 +97,14 @@ function getCredentialFromRushEnv(): VercelCredential {
  * @param teamId - The teamId can be provided to select the credential with the matching `teamId` from the
  * user credential store in `~/.rush-user/credentials.json`.
  *
- * @returns Credentials for Vercel Remote Cache
+ * @returns Credentials for Khulnasoft Remote Cache
  */
-export function getVercelCredentialTeamTokenPair(
+export function getKhulnasoftCredentialTeamTokenPair(
   teamId?: string,
-): Promise<VercelCredential> {
-  const fromVercelArtifactsEnv = getCredentialFromVercelEnv();
-  if (fromVercelArtifactsEnv.token) {
-    return Promise.resolve(fromVercelArtifactsEnv);
+): Promise<KhulnasoftCredential> {
+  const fromKhulnasoftArtifactsEnv = getCredentialFromKhulnasoftEnv();
+  if (fromKhulnasoftArtifactsEnv.token) {
+    return Promise.resolve(fromKhulnasoftArtifactsEnv);
   }
   const fromRushEnv = getCredentialFromRushEnv();
   if (fromRushEnv.token) {
@@ -122,13 +122,13 @@ export function getVercelCredentialTeamTokenPair(
  *
  * @param teamId - The teamId of the credential to set or update
  */
-export async function updateVercelCredentialForUserStore(
+export async function updateKhulnasoftCredentialForUserStore(
   token: string,
   teamId?: string,
 ): Promise<void> {
   if (/[^\w]+/.exec(token)) {
     throw new Error(
-      'The provided Vercel access token contains invalid characters.',
+      'The provided Khulnasoft access token contains invalid characters.',
     );
   }
 
@@ -162,7 +162,7 @@ export async function updateVercelCredentialForUserStore(
  *
  * @param teamId - The teamId of the credential to remove
  */
-export async function deleteVercelCredentialForUserStore(teamId?: string) {
+export async function deleteKhulnasoftCredentialForUserStore(teamId?: string) {
   const { globalCacheId, teamCacheId } = getCacheIds(teamId);
   if (teamCacheId) {
     await CredentialCache.usingAsync(