+++ packages/remote-rush/src/index.ts
@@ -1,4 +1,4 @@
-import { RushVercelRemoteCachePlugin } from './rush-vercel-remote-cache-plugin';
+import { RushKhulnasoftRemoteCachePlugin } from './rush-khulnasoft-remote-cache-plugin';
 
 // eslint-disable-next-line import/no-default-export
-export default RushVercelRemoteCachePlugin;
+export default RushKhulnasoftRemoteCachePlugin;