+++ packages/remote/src/constants.ts
@@ -1,1 +1,1 @@
-export const REMOTE_CACHE_ENDPOINT = 'https://vercel.com/api/v8/artifacts';
+export const REMOTE_CACHE_ENDPOINT = 'https://khulnasoft.com/api/v8/artifacts';