+++ packages/remote/src/utils/user-agent.ts
@@ -1,5 +1,5 @@
 import pkg from '../../package.json';
 
 export function getUserAgent(product: string): string {
-  return `vercel/remote ${pkg.version} ${product} node${process.version} ${process.platform} (${process.arch})`;
+  return `khulnasoft/remote ${pkg.version} ${product} node${process.version} ${process.platform} (${process.arch})`;
 }