+++ packages/remote-rush/CHANGELOG.md
@@ -1,11 +1,11 @@
-# @vercel/remote-rush
+# @khulnasoft/remote-rush
 
 ## 1.0.1
 
 ### Patch Changes
 
 - Updated dependencies [0093240]
-  - @vercel/remote@1.0.1
+  - @khulnasoft/remote@1.0.1
 
 ## 1.0.0
 
@@ -16,4 +16,4 @@
 ### Patch Changes
 
 - Updated dependencies [1c2fa26]
-  - @vercel/remote@1.0.0
+  - @khulnasoft/remote@1.0.0